+++ datarepository/src/main/java/com/library/repository/models/ShoppingInfoBean.java
@@ -19,18 +19,18 @@ public class ShoppingInfoBean {
      * sendTime : ["18:17","18:20","18:25","18:30","18:35","18:40","18:45","18:50","18:55","19:00","19:05","19:10","19:15","19:20","19:25","19:30","19:35","19:40","19:45","19:50","19:55","20:00","20:05","20:10","20:15","20:20","20:25","20:30","20:35","20:40","20:45","20:50","20:55","21:00","21:05","21:10","21:15","21:20","21:25","21:30","21:35","21:40","21:45","21:50","21:55","22:00","22:05","22:10","22:15","22:20","22:25","22:30","22:35","22:40","22:45","22:50","22:55","23:00","23:05","23:10","23:15","23:20","23:25","23:30","23:35","23:40","23:45","23:50","23:55","00:00","00:05"]
      */
 
-    private int totalPrice;
+    private long totalPrice;
     private int totalNum;
     private AddressInfoBean addressInfo;
     private ShopInfoBean shopInfo;
     private List<GoodsListBean> goodsList;
     private List<String> sendTime;
 
-    public int getTotalPrice() {
+    public long getTotalPrice() {
         return totalPrice;
     }
 
-    public void setTotalPrice(int totalPrice) {
+    public void setTotalPrice(long totalPrice) {
         this.totalPrice = totalPrice;
     }
